shuffle things around, extract pieces of sql code which can be reused
