+++ allennlp/semparse/contexts/__init__.py
@@ -1,2 +1,2 @@
 from allennlp.semparse.contexts.table_question_knowledge_graph import TableQuestionKnowledgeGraph
-from allennlp.semparse.contexts.sql_table_context import SqlTableContext
+from allennlp.semparse.contexts.atis_sql_table_context import AtisSqlTableContext
