+++ doc/api/allennlp.semparse.contexts.rst
@@ -21,7 +21,12 @@ allennlp.semparse.contexts
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: allennlp.semparse.contexts.sql_table_context
+.. automodule:: allennlp.semparse.contexts.atis_sql_table_context
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: allennlp.semparse.contexts.sql_table_context_utils
    :members:
    :undoc-members:
    :show-inheritance:
